move these out to lib folder
