+++ app/src/lib/editors/win32.ts
